+++ crates/ffi/src/convert/app.rs
@@ -1,0 +1,42 @@
+use crate::convert::string::{free_c_string, opt_string_to_c, string_to_c};
+use crate::types::AdAppInfo;
+use agent_desktop_core::node::AppInfo;
+use std::os::raw::c_char;
+use std::ptr;
+
+#[allow(dead_code)]
+pub(crate) fn app_info_to_c(a: &AppInfo) -> AdAppInfo {
+    AdAppInfo {
+        name: string_to_c(&a.name),
+        pid: a.pid,
+        bundle_id: opt_string_to_c(a.bundle_id.as_deref()),
+    }
+}
+
+#[allow(dead_code)]
+pub(crate) unsafe fn free_app_info_fields(a: &mut AdAppInfo) {
+    free_c_string(a.name as *mut c_char);
+    free_c_string(a.bundle_id as *mut c_char);
+    a.name = ptr::null();
+    a.bundle_id = ptr::null();
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::convert::string::c_to_str;
+
+    #[test]
+    fn test_app_info_roundtrip() {
+        let a = AppInfo {
+            name: "Finder".into(),
+            pid: 42,
+            bundle_id: Some("com.apple.finder".into()),
+        };
+        let c = app_info_to_c(&a);
+        assert_eq!(unsafe { c_to_str(c.name) }, Some("Finder"));
+        assert_eq!(unsafe { c_to_str(c.bundle_id) }, Some("com.apple.finder"));
+        let mut c = c;
+        unsafe { free_app_info_fields(&mut c) };
+    }
+}
